Improve StreamData.string/2
Before this commit, we would have skewed probabilities in some cases
when using "StreamData.string/2". For example, if called like

    StreamData.string([?a..?z, ?_])

then there would be equal chances for each character of being in ?a..?z
as well as being ?_ (resulting in a string full of underscores). With
this commit, we do a weighed frequency of characters so that in the
example above we would choose something in ?a..?z with probability 26/27
and ?_ with probability 1/27.