Add `max_run_time` configuration option
This allows users to set both a `max_runs` option to their property
tests, limiting the tests to a set number of generations, as well as a
`max_run_time` option to ensure that the tests don't take longer than a
certain time period. This is the first rough draft of this feature, and
I'm going to focus more on the documentation and some refactoring in
the next commit after I get a little feedback from folks. As it is now
it's functional, though.