Pass in options correctly in do keyword syntax.
Doing a:
  `check all i <- integer(), max_runs: 25, do: assert ...`

failed with a match error because the options were packed with the body.

Fixed by extracting the options for a second time with `&Enum.split/2`.
Match on the body again to make sure that people don't pass in options
after the body.
Pack the options in another list since compile_check_all expects them
that way.

Added a test with a counter to verify this behaviour.