Date generators (#218)
This is a reimplementation of the date generators by @Qqwy in #161. As such, it
is a (partial) implementation for #129.

The tests and documentation are entirely copy-pasted from his PR,
but the actual implementation is entirely new. As noted by @whatyouhide in a
comment to the original PR, I've opted to use the new `Date.add` function for
most of the cases.

I think the resulting implementation is a bit easier to follow. I have not added
Time, DateTime or NaiveDateTime, though I think we could probably do something
for those that follow this skeleton.

Co-authored-by: Qqwy / Marten <w-m@wmcode.nl>
Co-authored-by: Andrea Leopardi <an.leopardi@gmail.com>